+++ packages/plugin-hardhat/test/propose-upgrade-with-approval-beacon.js
@@ -27,6 +27,8 @@ test.beforeEach(async t => {
     './utils': {
       ...require('../dist/defender/utils'),
       getNetwork: () => t.context.fakeChainId,
+    },
+    './client': {
       getDeployClient: () => t.context.fakeDefenderClient,
     },
   }).makeProposeUpgradeWithApproval(hre);